backend/crewai/requirements.txt
crewai
crewai-tools
mem0ai
python-dotenv
